[ALPS09235115] blocktag: Use xarray for btag and mictx
[Description]
Replaced list-based blocktag management with xarray for better
performance and scalability. Updated related functions to use xarray
operations.

[Test]
bootup pass

MTK-Commit-Id: e3a72a77a114b5efbf935effb3f45612437aea6a

CR-Id: ALPS09235115
Feature: blocktag

+++ drivers/misc/mediatek/include/mt-plat/mtk_blocktag.h
@@ -27,7 +27,7 @@ enum mtk_btag_storage_type {
 
 #define BTAG_NAME_LEN           16
 struct mtk_btag_mictx_id {
-	__u16 id;
+	unsigned long id;
 	enum mtk_btag_storage_type storage;
 	char name[BTAG_NAME_LEN];
 };
